@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { SuggestionService } from 'src/app/services/suggestion/suggestion.service';
 import { FormSentComponent } from '../../components/form-sent/form-sent.component';
@@ -10,13 +10,13 @@ import { FormSentComponent } from '../../components/form-sent/form-sent.componen
   styleUrls: ['./suggestion.component.sass'],
 })
 export class SuggestionComponent implements OnInit {
-  suggestionForm = new FormGroup({
-    name: new FormControl(null, Validators.required),
-    email_address: new FormControl(null, [
+  suggestionForm = new UntypedFormGroup({
+    name: new UntypedFormControl(null, Validators.required),
+    email_address: new UntypedFormControl(null, [
       Validators.required,
       Validators.email,
     ]),
-    content: new FormControl(null, Validators.required),
+    content: new UntypedFormControl(null, Validators.required),
   });
 
   error: string | null = null;
